Fix for the bellow bug ========================================== amalgalite3.so: [BUG] Segmentation fault ruby 1.8.6 (2010-09-02) [i686-linux]
Aborted
========================================
In the above the binary was trying to find some database , but could not find it ,since it was in another directory, for which we do not have path for.
Amalgalite extension tries to raise an error but in the function am_bootstrap_cleanup_and_raise a char array, which is the error message, was tried to be freed so we get a Segmentation fault
Modified the am_bootstrap_cleanup_and_raise function to have additional parameter if the msg to be freed or not depending on the case if the msg is array or allocated pointer.