Fix incorrect loading/reload of schema [GH #16]
SQLite keeps the temporary and main meta information tables separate. The main
tables are in 'sqlite_master' and the temporary tables are in
'sqlite_temp_master'. Amalgalite was not finding the temp tables when looking up
the tables by name, and it was not reloading the schema correctly if there was a
change to the temporary table.